Move obfuscation & redaction into own package
