+++ data_masks/obfuscation_test.go
@@ -1,4 +1,4 @@
-package trace
+package data_masks
 
 import "testing"
