+++ data_masks/redaction_config.go
@@ -1,4 +1,4 @@
-package trace
+package data_masks
 
 import (
 	"embed"
