+++ trace/report_buffer.go
@@ -10,6 +10,7 @@ import (
 	"github.com/akitasoftware/akita-libs/client_telemetry"
 	"github.com/akitasoftware/go-utils/optionals"
 	"github.com/pkg/errors"
+	"github.com/postmanlabs/postman-insights-agent/data_masks"
 	"github.com/postmanlabs/postman-insights-agent/printer"
 	"github.com/postmanlabs/postman-insights-agent/rest"
 )
@@ -84,7 +85,7 @@ func (buf *reportBuffer) addWitness(w *witnessWithInfo) {
 			// The witness exceeds our per-witness storage limit. Obfuscate it to
 			// reduce its size while retaining its typing information.
 			printer.Debugf("Obfuscating oversized witness (%d MB) captured on interface %s\n", len(witnessReport.WitnessProto)/1_000_000, w.netInterface)
-			ObfuscateMethod(w.witness.GetMethod())
+			data_masks.ObfuscateMethod(w.witness.GetMethod())
 			witnessReport, err = w.toReport()
 			if err != nil {
 				printer.Warningf("Failed to convert obfuscated witness to report: %v\n", err)
